requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-dotenv==1.4.2
django-restframework==0.0.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Faker==15.1.1
idna==3.4
psycopg2==2.9.4
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
toposort==1.7
tzdata==2022.5
urllib3==1.26.12
